requirements.txt
asgiref==3.4.1
Django==3.2.9
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
